PDF interpreter - match observed Acrobat processing of Widget annots
No bug for this, arising from a customer file.

It appears Acrobat does not render Widget annotations unless the
annotation dictionary contains a /FT and a /T key, either directly in
the annotation dictionary or inherited from a Form Definition Field
dictionary (via the AcroForm /Fields entry).

This commit checks annotations to see if they are Widgets, then checks
to see if both /FT sand /T are present. If either is missing we walk
back up the FDF heirarchy using the /Parent key, storing any FT or T
keys in the annotation dictionary as we go.

We don't actually care what the associated values are, we only care if
they keys are present, so reversing the lookup is acceptable here.

We then re-test the annotation dictionary for the presence of both keys
and if either is missing we do not render the annotation.