PDF interpreter - Ignore invalid MediaBox array
Again, no bug, the specimen file is a confidential customer file.

The file contains, on page 60, an invalid MediaBox:

/MediaBox[ 0 0 612 792 0 0 720 917]

The 8 values were unloaded onto the stack, but only 4 were consumed
leading to errors later in processing. This commit checks the length
of the array and if it isn't exactly 4 then it exits and uses the
current page size instead.