PDF interpreter - do not unconditionally restore state in update_alpha
Bug #701644 "Error: /unregistered in --.pdfexectoken--"

The problem is caused by the insanely complicated (and inefficient)
way the file is constructed. We have a SMask in a Group, which is
enclosed within another Group.

When we call start a transparency mask we call pdf_prepare_drawing()
which calls pdf_update_alpha(). Crucially this is called after we have
entered a substream for an enclosing Group, but it does not begin a
new substream.

However, pdf_update_alpha() assumes that if the smask_id has changed
then a new viewer state has been saved, and the first thing it does
is restore that state. If no such state has been saved (as here) then
the code to restore the viewer state sees this as an error (which it is)

We fix this by checking to see if a viewer state has been pushed before
trying to restore to it.