PDF interpreter - handle ObjStm with contents which are non-delimited
No bug number, confidential file supplied by a customer.

This is in the same area as commit
d9889892feaac3ca09f85d9bbdc001b1aec50cc3

In this case the file is actually completely valid and exposes a flaw
in our processing of ObjStm streams. We 'tokenise' the entire contents
of the ObjStm, for performance reasons, when we first need to read an
object from it. This caused problems if we read too many objects which
was dealt with in the commit above.

In this case, however, what we have is objects which do not have any
delimiter, so the tokeniser carries straight on, instead of stopping
when it reaches the end of the object (as given by the difference
between successive offsets). For example, if we have three null objects
then the (decompressed) stream would read nullnullnull, each object
has a different offset and the length of each is given by subtracting
one offset from the next.

What we were reading was, literally, nullnullnull which of course is not
legal.

This commit leverages off the code in the commit to fix the extra
objects and uses the same approach, we set up a ReusableStream on the
decompressed ObjStm so we can seek around in it, then read each object
by seeking to the offset, setting up a SubFileDecode (so we stop at the
end of the object) and reading the object.

That works as well for this case, albeit with a (negligible in this
case) performance penalty). We check at the end of reading the objects
individually, and if we don't have enough objects then we raise a
warning.