Update SubFileDecode to work with a length as well as a string
The SubFileDecode filter was only half-implemented, while we used the
EOD string and length, we did not use EODCount, which meant we always
stopped on the first occurence of EOD and we could not specify just the
length of the stream by setting EODCount but with a string length of 0.


Fix pdfi_read_bytes so that any buffered data is returned even at EOF

We tested the stream EOF before returning any buffered data which could
lead to us losing any data which had been returned to the stream
buffer by pdf_unread_bytes(). We now check that unread_size is 0
and return any unread bytes even at EOF. We don't return 0 until we
have reached EOF and there are no unread bytes.