Use SubFileDecode to limit bytes read from an ObjStm
Inspired by a customer file.... An object in an ObjStm is not delimited
with x y obj...endobj, its a naked object. Some object types are
self-delimiting such as dictionaries and arrays, but other types are
not.

In particular null objects and numbers are not delimited. Because the
objects in an ObjStm need not have any white space, this can lead to
errors if we are try to read tokens. Consider two number 123 and 456
which are stored consecutively in an ObjStm, the offset array would
contain something like x 0 y 3, so that the offset of object x is 0,
and the offset of object y is 3, but the actual decompressed stream
content would be '123456'. We have to know that the initial object
contains only 3 bytes and stop after that.

We were not previously doing this.

This commit adds a SubFileDecode around the entire ObjStm stream object
using the /Length key to lilmit the number of bytes which can be read.
This is because the offsets array only contains offsets to each object
and so cannot be used to find the length of the last object in the
stream.

When reading an object, except for the last one in the stream, we apply
another SubFileDeocde to the compressed stream to limit the amount of
uncompressed data we read to the difference between the offsets of
two consecutive objects.

This prevents us running past the end of an object, even if it is not
delimited in the stream.