jbig2dec: Fix memento interfering with allocator callback names.
Memento changes the names of the callbacks when they are set, but
not when the structure is defined in the command line tool. Work
around this by including memento prior to the structure being
defined so that memento changes the names in both places.

The alternative would be changing the jbig2dec API yet again.