Fontmap: cope better with bare filename maps
We don't automatically add FontResourceDir to the libfile search path, so a
name to plain file name mapping will only work if the directory containing the
file is explicitly added to the search path. For example:

/Dingbats		(d050000l.pfb) ;

With this change, in addition to the search path and the bare file name, we'll
also try opening the file in the FontResourceDir, so users can use such custom
mappings simply by adding the font files to the FontResourceDir (usually:
Resource/Font).