Make the 'i' setflat operator match gs code
Spec says the value is 1-100, with 0 meaning
use the device default.

But the gs implementation does a "min(val, 1)" in the ps code
and gs_gstate_setflat will pin it to >= 0.2, so basically we get
a range of 0.2-1.0 instead.

This now matches what Adobe, gs, and mupdf do.  See Bug 555657.