Bug #699331: better support for cross-compiling
This is mainly moving from using uname to identify the the target platform,
to using the 'host' value to identify the target platform, and 'host' and
'build' values to know when we're cross-compiling.

There are also a couple of tweaks related to cross-compiling to MSYS.

Also, make double sure API symbols are always visible. This affects builds with
gcc and compatible compilers.

Remove use of $cross_compiling variable

Credit to djcj for most of the changes