Pattern handling- relax a heuristic
Bug #701753 "Error reading a content stream" with pdf using Type3 font"

This file uses a type 3 font with a FontMatrix more usually found in
type 1 font; [0.001 0 0 0.001 0 0]. It is also a coloured font, and
uses a Pattern colour space with a Shading function to colour the text.

The Pattern cell, when scaled by the Font Matrix, results in a mapping
of the step matrix to device space where the matrix determinant is less
than 1/1,000,000. We detect this as a degenerate matrix and throw an
error.

Acrobat renders an empty page.

So relax this heuristic further from 10^-6 to 10^-9, this allows the
file to render and doesn't cause any changes in our regression tests.

Of course, at this cell size, the pattern renders blank, just as it
does with Acrobat.