Use caller defined pkg-config and cups-config unconditionally
In an effort to better handle cross-compilation, we use a check combining
AC_PATH_TOOL() and AC_PATH_PROG() - where AC_CHECK_TOOL() will first attempt to
find the requested tool *with* a target-triplet prefix (for example,
"x86_64-linux-gnu-pkg-config") and if that fails, it will fall back to the
unadorned tool (e.g. "pkg-config"). If we are cross-compiling, and both
AC_PATH_TOOL() and AC_PATH_PROG() return the same value, we assume that
AC_PATH_TOOL() has fallen back to the default, non-cross-compile version, and we
disable using the given tool.

We do this for both pkg-config and cups-config.

This does not work well for every cross-compile environment, however, since not
all use the convention of the target triplet prefix.

We already allowed the caller to specify both a pkg-config and cups-config to
use, but we still applied the above outlined test.

Now, when the tool is explicitly specified, do *not* do the
AC_PATH_TOOL()/AC_PATH_PROG(), assume the caller knows what they are doing, and
just go ahead and use it.

Additionally, for cups-config, if the caller explicitly specified a cups-config,
and/or explicitly set --enable-cups then make not finding the cups libraries a
configure hard fail. By default, we'll still just warn, and exclude the cups
devices.