Update genfontmap.ps for SAFER default
Bug #701766 " genfontmap.ps Unrecoverable error, exit code 1"

The genfontmap.ps program, supplied as part of the support tools in
'toolbin' rather than Ghostscript proper, relies on the non-standard
PostScript extension operator '.sort'.

As part of the security hardening, as many non-standard operators as
possible were removed from Ghostscript when SAFER is set, including
.sort, and with version 9.50 SAFER becomes the default.

This commit simply updates the genfontmap.ps program to check that it
is being run with NOSAFER, emits a warning and exits if this is not the
case. Also update the comments to indicate why we are doing this.