Memento tweaks: libbacktrace warning/instructions.
If we can't load libbacktrace.so, give a warning that backtraces
will be cryptic.

Add info to memento.h about where people can get libbacktrace
from, how to build it, and where to install it.

Also add some code from Julian Smith to call gdb to extract
backtrace data; this doesn't work for me, so it's disabled
by default for now.