Prevent a heap-buffer overrun
Bug #701791 "global-buffer-overflow at devices/gdevpjet.c:177 in pj_common_print_page"

We were not setting the unused bytes at the end of a line to 0, which
later caused us to read uninitialised data from the line, and attempt
to use that as the index into an 8-byte table. If the uninitialised
data was greater than 16 then we would read off the end of the table.

Its 'probably' benign since we aren't using tis data, and we probably
won't try to read from an invalid address, but its poor practice and
its easily fixed. Setting the buffer to 0x00 before we start ensures
that any padding bytes are in the valid range for the table.