Color Laserjet device - properly check params
Bug #701796 "Segmentation fault at devices/gdevclj.c:269 in clj_media_size"

The test in clj_media_size() as incorrect, it didn't signal an error if
there was no HWResolution in the param list. This left the parameter
'fres' undefined leading to a potential invalid memory read, or a
divide-by-zero error.

This commit returns an error if we get a param list with no HWResolution
I'm not absolutely certain if this is correct, but at least signalling
an error is better than a seg fault.