Add convenience routines for array and dict, refactor code
Add routines for inserting things in arrays and dicts.  Refactor some
of the old code that put things in dicts.

Basically now the name will be created in the routine, instead of by
the caller.  Things can be inserted into a dict by specifying the
key as a string instead of having to create an object in the caller.

Some of these new routines will be used by image code that is WIP, or
at least that's why they were written...