Add support for SMaskInData = 2 and some cleanups
SMaskInData = 2 is in fts_17_1718.pdf, but supporting it
properly actually doesn't affect that sample.  Not sure what that
means.

Anyway, do same as the gs code, which is to put a "Matte" array
of size=comps, full of 0's.

Re-arrange code in main pdfi_do_image() routine so that 'comps' will
be available at the time the pdfi_make_smask() routine is called.

Also cleaned up some initialization and error handling.