Fix bugs 701787, 701806, 701810. Problems with cdj970 device.
As Robin mentioned in bug 701787, the device was changing resolution AFTER
the device had been opened and all of the buffers configured.

Move the "one time" initial setup into the open function, leaving the code
to write the header in the print_page function. Presumably that only needs
to be written once even if there are multiple pages.

Also add a check for valid resolutions since it appears that the intent was
to have the "Quality" parameter set up 300 or 600 dpi. Other deskjet devices
have this type of check.

Add a gs_closedevice if the Quality is changed since this will change the
resolution and thus the page buffer geometry.

Lastly, fix cdj970_put_params so that errors are not ignored for all but the
last (which happened to be BlackCorrect).

These changes prevent the bugs cited, but remain untested except for some
parameter testing to make sure bad values don't cause memory violations. It
does seem that some parameter values that are out of range (like Quality) are
ignored, but that may be intentional.