PDF interpreter - preserve colour space on Image XObjects with Do
No bug number the customer file is large and looks confidential.

The problem is that the PDF file sets a custom colour space, then
draws an image, the image is in a different colour space (DeviceRGB),
the file then sets a new colour in the original space.

The PDF interpreter was changing the colour space in order to draw
the image, but was not resetting it afterwards, which meant that the
attempt to set the colour failed as the colour space was inappropriate.

I'm *amazed* this hasn't arisen before, its so clearly wrong.

This commit gets the colour space before executing the image, and puts
it back again afterwards. We have to be careful to do the restoration
in a stopped context because this code can be called when drawing a
glyph, and we are not permitted to change colour space while in an
uncoloured glyph, it throws an error.

In addition, the appearance generation for /Square annotations (and
other types) was setting the colour space to white and filling the
annotation rectangle, even if the /IC (Interior Color) key was not
present. Acrobat appears not to do this so I've modified the code to
not fill the annotation if there is no /IC in the annotation dictionary.