Old Type 1 hinter: protect against signed overflow
To avoid overflow during hinting, we tweak the balance of the size of the
coordinates and the scaling in the hinter (this is only dealing with coords
so large that hinting is largely pointless anyway).

But the code that does that didn't protect against signed overflow in the
starting coordinates.

This is a naive protection to prevent an infinite (or near so) loop, but
a) it's old code that shouldn't be in use anymore, and b) in practice it will
make no difference to rendering with coordinates that large.