Add code to build a spot function from a Halftone spot
Apart from the actual function, the spots are all the same so we can
use common code to generate the function.

make pdfi_build_sub_function actually set the error code on an error
condition, before returning, so we know if this fails and don't try
to use broken functions. Also pdfi_build_function_0()