Add some halftone handling
This only handles type 1 halftones at present, but it does it using
the graphics library 'multiple' halftone instead of using the 'screen'
functions (disabled code with #if 0) which only work with a single
colour.

This is so that we can use the same code (as Ghostscript does currently)
for single and multiple colour halftones.

This is far from complete, we still need to handle threshold arrays
(type 6, 10 and 16) and multiple colour halftones (type 5) but we can't
do type 5 until the others are in place.