Improve handling of microscopic fonts with text rendering mode 3
Bug #701875 "empty TIFF-file is generated when using -r300, correct output with -r72"

In the past we've seen PDF producers set the text rendering mode to 3
and then (goodness knows why) set a CTM where one of the scale factors
is 0. In this case the scale factor isn't 0 but its absolutely minute,
5.3x10-6.

This evades our existing code, so here we improve it; we check to see
if the CTM scales in either direction by a really small amount instead
of 0 (0.000001).

In addition, we previously tried to patch up the CTM and use it anyway
but really this is only going to make sense is one scale factor is
tiny (or 0) and the other is sensible. If that's not the case then
any displacement due to drawing the text will also be tiny. So lets
simply ignore it.