Clear out gs_gstate at the start of clist playback.
There are some elements of the (stack based) gs_gstate that were not being
set by either GS_STATE_INIT_VALUES_CLIST, or gs_gstate_initialize, so set
the entire structure to 0 before setting the values we care about. This
was prompted by a customer that has made modifications to the clist playback
that relied on gs_gstate.color[0/1].color_space being NULL or valid, but
since it is so quick and only happens once per band, go ahead and do it.