Fix devices that ignore return codes from gdev_prn get_bits & copy_scan_lines
Discovered these when working on bug 701845 'devicen' device. Checking
uncovered a plethora of other places that used one of these functions
assuming success.

Also the gdev_prn_copy_scan_lines had places that assumed the return code
was a valid line_count, so if negative, could memset many lines in front of
an allocated buffer.

The code in a few of these usages is strange, and we don't have a way to
check them, so this may not match the original intent.

Also, while perusing all of these files, check devices for alloc's that
assume success and fix them (usually return gs_error_VMerror).