pdfwrite - For PDF input, copy any UserUnit to the output
If the input is PDF, and the input file uses the ugly UserUnit hack to
work around MediaBox limits in Acrobat, and the CompatibilityLevel of
the output is at least PDF 1.6, then do not apply the UserUnit to the
MediaBox and do not scale the content. Instead pass the value of
UserUnit to the pdfwrite device and insert it into the output page.

We use a special_op to inquire whether the device would like us to pass
it the UserUnit. The pdfwrite device checks the current PDF level
being emitted and either returns true, or false if too low a level is
requested.

If it returns true the PDF interpreter does not apply the UserUnit
scaling but instead uses a second special_op to inform the device what
of the value of UserUnit.

The pdfwrite device adds any non-unity value of UserUnit to the page
dictionary when emitting it.

If -dNoUserUnit is set we neither apply the UserUnit scaling nor pass it
to the device.