PDF interpreter - don't ignore SUBSTFONT
The PDF interpreter attempts to find a 'better' match for missing fonts
by using a list of TrueType mappings (eg ArialMT, CourierNew) and a
list of mappings (gs_font.ps, /.substitutefaces)

Failing to find a match there it attempts to use the FontDescriptor
Flags if present. As far as I can see that basically means we use
Times-Roman for serif fonts and Helvetica for sans serif.

This commit checks to see if SUBSTFONT is set. If it is we ignore all
this nonsense and simply use the substitute font the user told us to.
Otherwise we continue as before.

The diff is actually very simple, but the routine was very large so
I decided to move the actual substitution code out into a separate
function, to simplify the logic in pdffindfont which makes the
differences appear substantial. All that has really changed is that the
code to do the substitution, after checking for the existence of a
FontDescriptor, checks to see if SUBSTFONT is defined. If it is we
behave as if the font had no FontDescriptor and use the normal font
substitution machinery.