Bug 701903: (ps2write) clean up image enumerator streams.
At the end of an image (especially if the "end" is because of an error
condition) make sure we clean up any streams created for the handling of the
image data.

This saves leaving the streams around for gc to cleanup, and avoids a problem
where a special purpose stream (for color images) holds a reference to a
color space and graphics state either or both of which could be freed by a
Postscript restore before the garbager gets called to deal with unreferenced
objects.