Support CIDFSubstFont being a path/file
Preivously, to use a custom CIDFont substitute in a custom location, it was
necessary to both both CIDFSubstPath (for the location) and CIDFSubstFont
(the file name).

We'll now check the CIDFSubstFont value and if it looks like it includes a path
(i.e. contains a directory/file separator string) then we'll use it as is,
ignoring the CIDFSubstPath