Attempted fix for Coverity ID 350194
Thanks to Robin for pointing me at the right bit of code.

There was a pointer cast and address of (&) on an array which was
already a pointer. This is, of course, silly, so remove the cast and
&.

At the same time, the quad_glyph_list was using signed shorts to hold
the Unicode points, when it should have been using unsigned shorts, so
fix that here too.