Coverity ID 350216 - prevent dangling pointer
The variable pdcs is declared an immediately initialised to be
dsc->dcs2. Coverity correctly points out that if dsc->page_count is 1
we enter dsc_alloc_string, which can call dsc_reset() where,
if dsc->dcs2 is set, it will be freed and dsc->dcs2 will be set to
Null, leaving pdcs dangling.

By deferring the assignment of pdcs until it is actuallused we can
be certain that it is valid.