txtwrite - fix buffer overflow
Bug #701877 "heap-buffer-overflow at devices/vector/gdevtxtw.c:2114 in txt_add_fragment"

We were trying to read too many entries from the enumerator 'Widths'
array, because we were reading the number of Unicode code points,
instead of the number of character codes.

At the same time.....

re-instate the code which uses & on the glyph lists and casts them to
appropriate pointers. While gcc happily works with the code that was
modified to make Coverity happy, Visual Studio absolutely does not.
In order to make Visual Studio happy we do need the crazy pointers and
casting. I think this is wrong, but this way works on both compilers
and I couldn't find any other construction which did.