Fix the Lexmark contributed device
Bug #701905 "heap-buffer-overflow at contrib/gdevlx32.c:1509 in convbuf"

The problem here is actually much earlier. In encode_bw_buf() there is a
comment (line 1771) about starting before the first non-white pixels
in order to "give the head a little more room to accelerate properly".

If the first non-white pixel is less than 3 pixels from the left edge
of the raster, this will cause the calculation of 'left' to become
negative. It seems to me this will immediately corrupt the data. In
addition right is calculated in a similar fashion, if right is also
close to the right edge of the raster, then we can end up with numcols
being more than the width of the raster.

This argument is passed to convbuf() and that tries to read off the end
of the input data, because numcols exceeds the bytes in the raster.

I'm pretty dubious about this whole piece of code but since I don't
have a printer to test with I'm reluctant to make major changes. This
commit simply clamps left to 0 and right to the right edge of the
raster. The comment implies that the extra spacing probably isn't useful
and certainly suggests that removing it won't be harmful.