Bug 701660  Trans. Text knockout missing ET
The file has contents that look like this

BT
 stuff
BT
 stuff
ET

more stuff

The first BT meets the conditions to push the text group.
Unfortunately it is missing the ET, so the group is never popped.
These fixes will make sure that if we are in a text group and
encounter another BT we popped the current text group.