Ensure paths from Fontmap end in a directory separator
Since it's normal for multiple font files to be grouped together in a directory,
we strip the file name off, and add just the directory to the permit read list,
rather than flood the list with every font file.

The problem is, we weren't appending the directory separator character showing
the directory was reading permitted.

This commit does that.

Also, add extra detail to the documentation about this aspect of file
permissions lists.