Bug 701625  Treat CMY color space as additive
Our support of CMY devices is currently not ideal with
respect to proper color management and the transparency
imaging model as described in Bug 697965.  My plan is
to make a psd cmy based device and get things working
properly in terms of spot colors and ICC color management.
This will take a bit of effort and likely have a bug
tail.  For now for Bug 701625 an easy solution is to
treat the color space as additive.  In this case, the
output looks reasonable.