Fix coverity #351085; toff_t's are unsigned.
The tiff lib uses a 'seek' function that apes 'fseek' in that
it takes a whence field. In order for the whence field to be
useful for 'SEEK_SET' (and half the cases of SEEK_CURR), it
needs to be able to take negative offsets. The tiff seek
function only passes positive offsets though as it uses an
unsigned type.

It is therefore pointless to test against 0.