Fix indeterminism in gdevatx.c
The compression used by this device relies on compressing pairs of
bytes; as such we have to be careful that when we offset to the end
of the line we don't move inot uninitialised data.

We fix this here by blanking an extra byte, and being smarter about
where we start the search for trailing empties from.

This should resolve the differences we see in the all-devs test.