Fix indeterminism in gdevp201.c
The printers herein read 1bpp data from gs, and then compressed
in terms of bytes. For the case where w%8 != 0, we'd have
uninitialised bits. Solve this by masking the last byte of each
line as appropriate.

Also, ensure that we don't 'overread' lines, check the return
code from the get_bits call, and blank any lines of the buffer
required to bring us up to a multiple of the stripe height.