Coverity 350201, 350191: Add some notes to the code.
Coverity spots that pdf14_fill_path checks for ppath being
NULL in one branch, hence assumes that it can be NULL in all
cases. Furthermore it spots that it is passed into
gx_default_fill_path, which can in some circumstances, dereference
it without checking it first.

The reason for this is that fill_path can permissibly be called
with a NULL path if we want to fill the given clipping path. This
is only used for shadings and patterns, which is exactly the case
checked for within gx_default_fill_path and pdf14_fill_path.

We'll resolve this with a 'false positive' in Coverity, but have
added the comments for the benefit of future readers.

Coverity spots the same thing in pdf14_stroke_path, but there it
really makes no sense for ppath to be NULL, so just eliminate the
check.

Credit to Julian Smith for the investigation on this.