Update Freetype to 2.4.10
Plus makefile updates to support it

Bug 701288: Handle SEAC ligature glyph metrics correctly.

As originally intended, a Type 1 SEAC charstring would be used for an accented
glyph (like an a-caron, or u-umlaut) where the advance with of the SEAC glyph
is the same as that of the 'base' glyph ('a' or 'u').

It this case, it is not uncommon for the SEAC to skip using an (H)SBW op code
and rely on the one from the base glyph.

But out of spec fonts also use SEAC glyphs ligature glyphs (like 'oe' or 'fi'),
and in those cases the overall advance width is greater than that of the 'base'
glyph.

In this case, we have to allow for the SEAC glyph having included an (H)SBW and
if it has, retain those values, rather than those from the base glyph.

Rename file and make file change to match

Avoids file name clash with freetype 2.4.10

Work around a change in the zlib API for 1.2.11

where it's used in the Freetype/zlib interface debugging code.