Fix problems seen with m8510 device and improve gdev_prn_copy_scan_lines
Previous patch made this device pay attention to return code from the
gdev_prn_copy_scan_lines, which caused the all_devices test to get an error.

Also this device would get an ioerror if the page width was not a multiple
of 8. Fix m8510_output_run to round up when setting out_end, bullet proof
check of count so we don't try and print with a negative count.

Also fix gdev_prn_copy_scan_lines to return count == 0 if the starting
line is past the end of the page, and fill lines past end of page with
zeroes if the count did not fill it (or if there was an error from
gdev_prn_get_bits).