Initial effort towards proper page group color handling
The issue this branch is trying to solve is to ensure
that the alpha blending occurs in the proper page group
color space.  If the page group is CMYK and the device is
RGB then the final alpha blend must occur prior to the
color conversion.  Currently with the head code this is
not the case.