Change some default values in the context, particularly 'printed'
The context gets initialized to 0, but some of the flags should actually
default to 'true'.  This fixes the ones I know about.

Note that for 'printed', it should actually default to 'true' if
OutputFile is set, and 'false' if it's not. But this means it is
'true' basically almost all the time, so I just set it to 'true' for
now until I can figure out how to test OutputFile at that point in the
flow.

This means some annotations won't render unless '-dPrinted=false' is specified
on the command line, but it should match the behavior of gs in these cases.