Create devdevs.tr correctly
devdevs.tr is an interim list of output and core library "devs" that gets
created and then used, ultimately, to create the gconfig.h header file
(which has the bulid configuration of the graphics library), and the linker
scripts.

When devdevs.tr required creation, we were always appending to the file (if it
already existed), not overwriting it, meaning the list would grow indefinitely,
and contain duplicates.

This looks like a typo: the first line of devdevs.tr creation used a '-a'
echogs paramater, rather than the '-w' parameter.

I think this should solve the problem.