PDF interpreter - improve /Square annotation appearance creation
Bug #701889 "Annotation objects missing or incorrectly interpreted when appearance stream missing"

First of probably several changes. In this case we were failing to
stroke the border of a /Square annotation if it had no interior colour
(we did not render the annotation at all).

In addition, when element 3 of the Border array was 0 we were still
drawing a Border, which we should not do (this didn't show up until we
fixed the problem above).