PDF interpreter - generate callout (/CL) lines for FreeText annots
Bug #701889 "Annotation objects missing or incorrectly interpreted when appearance is missing"

FreeText annotations can have a /CL (callout) entry, which defines a
line to be drawn to the point referenced by the text.

This is a first pass at the problem, Acrobat generates arrow heads for
the callout lines, this patch does not.