PDF interpreter - cope with legal but sub-optimal indirect /Pattern
No bug, the report is from a customer who wants the file to be treated
confidentially.

The files uses a Pattern colour space, which it defines in a highly
peculiar fashion:

stream
....
/CS0 cs /P0  scn

....
endstream

CS0 is defined as:
  /CS0 22 0 R

and object 22 is:

22 0 obj
[ 23 0 R ]
endobj

23 0 obj
/Pattern
endobj

The PDF interpreter wasn't prepared to handle the name being an indirect
reference.