PDF interpreter - Accept /ColorSpace as well as /CS in group attribute dictionary
Again, no bug report, the customer has asked that the file be kept
confidential

The problem is due to a group attribute dictionary for an SMask. The
dictionary is required to contain a /CS (colour space) entry in order
to correctly set the background. In this case, however, the dictionary
contains a /ColorSpace entry.

I've chosen to implement treating the /ColorSpace as a /CS entry, this
is not what Acrobat does, but mimicking Acrobat's behaviour is more or
less impossible for us. We do emit a warning.