Fix  fill_stroke SegFault caused by stroke pattern being ejected from cache
Add a method to allow a tile to be "locked" in the pattern_cache so that
do_fill_stroke can signal the cache logic to not free that tile.
This also requires the 'is_locked' flag to be passed through the clist,
and that the stroke color (which is locked) is written before the fill
color.