txtwrite - don't close the device on media size changes
Bug #701971 "txtwrite incomplete for certain files and differs when stdout versus file is output"

The problem is nothing to do with stdout vs file. The difference is due
to the fact that the input PDF file has multiple pages with different
media sizes. When the media size is changed the default put_params()
method closes and reopens the device. When the txtwrite device is
closed it closes its output file. Re-opening the device opens a new
file, if the filename doesn not have a %d then the sam file is opened
which will of course truncate it.

Writing to stdout obviously doesn't have this behaviour when the output
file is closed.

We don't need to close the txtwrite device when the media size changes
so here we patch the 'is_open' flag to prevent the default method
from closing the device.