Add funcs to deal with stream Length, and use them.
This adds a Length to the pdf_dict type, and since dicts are created
in various ways I decided to do a lazy evaluation, which means we also
need to keep track of whether Length is valid.

The is_stream flag indicates whether there was a Length key or not.

A Length of 0 means either there was no Length key, or it's a
zero-length stream, which either way the user can probably make sense
of.

Also cleaned up error handling in pdfi_create_iccbased() which would have
had a memory leak.

Also fixed bug in pdfi_build_function_0() where 'code' was being checked
instead of 'Length'.  Would have only mattered if there was an error.