PDF interpreter - apply RD array to Circle, Square and FreeText, apply Rotate to FreeText
Bug #701889 "Annotation objects missing or incorrectly interpreted when appearance stream missing"

The /RD array makes certain kinds of annotations be drawn in a smaller
area. Although the spec does not mention it, Acrobat applies the
/Rotate key if found to the text in a FreeText annotation.

Still more to do on the text; we are nto fitting the text correctly
when Rotate is 270, not certain why. Also we are not drawing Borders
when we should, some text is missing and text is in the wrong colour
(due to not applying /DS)